Migrated to null safety

+++ example/lib/main.dart
@@ -24,9 +24,9 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   final cropKey = GlobalKey<CropState>();
-  File _file;
-  File _sample;
-  File _lastCropped;
+  File? _file;
+  File? _sample;
+  File? _lastCropped;
 
   @override
   void dispose() {
@@ -58,7 +58,12 @@ class _MyAppState extends State<MyApp> {
     return Column(
       children: <Widget>[
         Expanded(
-          child: Crop.file(_sample, key: cropKey),
+          child: _sample == null
+              ? SizedBox()
+              : Crop.file(
+                  _sample!,
+                  key: cropKey,
+                ),
         ),
         Container(
           padding: const EdgeInsets.only(top: 20.0),
@@ -69,10 +74,6 @@ class _MyAppState extends State<MyApp> {
               TextButton(
                 child: Text(
                   'Crop Image',
-                  style: Theme.of(context)
-                      .textTheme
-                      .button
-                      .copyWith(color: Colors.white),
                 ),
                 onPressed: () => _cropImage(),
               ),
@@ -88,18 +89,19 @@ class _MyAppState extends State<MyApp> {
     return TextButton(
       child: Text(
         'Open Image',
-        style: Theme.of(context).textTheme.button.copyWith(color: Colors.white),
       ),
       onPressed: () => _openImage(),
     );
   }
 
   Future<void> _openImage() async {
-    final pickedFile = await ImagePicker().getImage(source: ImageSource.gallery);
+    final pickedFile =
+        await ImagePicker().getImage(source: ImageSource.gallery);
+    if (pickedFile == null) return;
     final file = File(pickedFile.path);
     final sample = await ImageCrop.sampleImage(
       file: file,
-      preferredSize: context.size.longestSide.ceil(),
+      preferredSize: context.size?.longestSide.ceil(),
     );
 
     _sample?.delete();
@@ -112,9 +114,9 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future<void> _cropImage() async {
-    final scale = cropKey.currentState.scale;
-    final area = cropKey.currentState.area;
-    if (area == null) {
+    final scale = cropKey.currentState?.scale ?? 1;
+    final area = cropKey.currentState?.area;
+    if (area == null || _file == null) {
       // cannot crop, widget is not setup
       return;
     }
@@ -122,7 +124,7 @@ class _MyAppState extends State<MyApp> {
     // scale up to use maximum possible number of pixels
     // this will sample image in higher resolution to make cropped image larger
     final sample = await ImageCrop.sampleImage(
-      file: _file,
+      file: _file!,
       preferredSize: (2000 / scale).round(),
     );
